requirements-server.txt
flask==3.0.0
gunicorn==21.2.0
numpy==1.26.4
tensorflow==2.20.0
tensorflow-io-gcs-filesystem==0.37.1
tfswin==4.1.1
pillow==11.3.0
requests==2.32.5
python-multipart==0.0.6
